Regulation Model
- Abstracted `Describable` concern from DescribableObject
- Regulation is `InputModel` which is `Describable`
