+++ lib/law.rb
@@ -7,6 +7,8 @@ require "spicerack"
 
 require "law/version"
 
+require "law/concerns/describable"
+
 require "law/describable_object"
 require "law/permission_base"
 require "law/regulation_base"
