+++ lib/law/regulation_base.rb
@@ -4,7 +4,9 @@ require_relative "regulations/laws"
 
 # A **Regulation** is the "lock" which has a matching **Permission** "key".
 module Law
-  class RegulationBase < DescribableObject
+  class RegulationBase < Spicerack::InputModel
+    include Describable
+
     include Regulations::Laws
 
     type_name "Regulation"
