+++ spec/law/describable_object_spec.rb
@@ -5,6 +5,8 @@ RSpec.describe Law::DescribableObject, type: :object do
 
   it { is_expected.to inherit_from Spicerack::RootObject }
 
+  it { is_expected.to include_module Law::Describable }
+
   it_behaves_like "a describable object" do
     let(:example_class) { example_describable_class }
   end
