+++ spec/law/regulation_base_spec.rb
@@ -3,7 +3,9 @@
 RSpec.describe Law::RegulationBase, type: :regulation do
   include_context "with an example regulation"
 
-  it { is_expected.to inherit_from Law::DescribableObject }
+  it { is_expected.to inherit_from Spicerack::InputModel }
+
+  it { is_expected.to include_module Law::Describable }
 
   it { is_expected.to include_module Law::Regulations::Laws }
